First commit exploring a different design
Instead of returning simple bool and error, we will return a slice of
PkgInfo which can then serve for matching in pkg package. This needs a
further redisgn due to some redundancies in the code, but offers more
than just matching possibilities. We gain a list of all or queried
packages.