This is an attempt to refactor redesigned API
This commit ships a massive overhaul of API. It changes the method names
which are more straightforward semantically. It tries to abstract a
generic package manager better - this might introduce a bit of pain at
some point in the future but for the moment they should represent a
decent model.